requirements.txt
# FastAPI va asosiy kutubxonalar
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Vector Store
qdrant-client==1.7.0

# Word fayl bilan ishlash
python-docx==1.1.0

# OpenAI 
openai==1.54.3

# Yordamchi kutubxonalar
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3

# Monitoring va logging
prometheus-client==0.19.0